openstack-redis-pronet-v1.3/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
flask-dance[google]==7.0.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
openstacksdk==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.2